+++ include/Tools/LineTool.h
@@ -15,8 +15,8 @@ public:
 	void OnToolChanged();
 	void OnSpaceClicked(const ImVec2& position, const ImVec2& snapped, bool left, bool right, bool wheel);
 	void OnShapeClicked(const ImVec2& position, const ImVec2& snapped, bool left, bool right, bool wheel, ShapeID shape);
-	void OnMouseHovered(const ImVec2& position, const ImVec2& snapped, float dx, float dy);
-	void OnMouseDragged(const ImVec2& position, const ImVec2& snapped, float dx, float dy);
+	void OnMouseHovered(const ImVec2& position, const ImVec2& snapped, int dx, int dy);
+	void OnMouseDragged(const ImVec2& position, const ImVec2& snapped, int dx, int dy);
 	void OnMouseReleased(const ImVec2& position, bool left, bool right, bool wheel);
 	void OnLayerSelected(LayerID layer);
 	void CancelShape();
